requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.3.0
Django==1.11.15
django-filter==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.8.2
idna==2.8
jedi==0.13.1
Markdown==3.0.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.6
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
requests==2.22.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.25.2
wcwidth==0.1.7
